+++ layers/lang/rust.lua
@@ -5,7 +5,7 @@ local L = { name = "Rust Language" }
 
 local function setup_dap()
 	local dap = require("dap")
-	dap.adapters.codelldb = require("LYRD.configs.codelldb")
+	dap.adapters.codelldb = require("LYRD.shared.dap.codelldb")
 	local debug_configuration = {
 		{
 			name = "runit",
